The delete_net_address function was deleting too much by default
- This function is called when a record is removed from the
net_dns_entries table.  Originally, it was removing an associated
address field when any DNS alias was deleted.  Unfortunately,
this had the unforseen consequence of deleting all DNS aliases
associated with the address, even if what the user meant to do
was only delete one of the CNAMES.  I updated the function, so
it only deletes the associated address if the user is requesting
the authoritative alias is deleted.

+++ schema/schema.sql
@@ -86,7 +86,7 @@ $$ language plpgsql;
 create or replace function delete_net_address() returns trigger as
 $$
 begin
-delete from net_addresses na where na.net_address_id = old.net_address_id;
+delete from net_addresses na where na.net_address_id = old.net_address_id and old.authoritative='t';
 return old;
 end;
 $$ language plpgsql;
